Move path implementation to visitor internal package, move record visitor to a internal record package so that we will have room to provide a schema visitor later
