+++ daikon/src/main/java/org/talend/daikon/avro/visitor/record/RecordVisitor.java
@@ -10,7 +10,7 @@
 // 9 rue Pages 92150 Suresnes, France
 //
 // ============================================================================
-package org.talend.daikon.avro.visitor;
+package org.talend.daikon.avro.visitor.record;
 
 /**
  * An Avro record visitor
