+++ daikon/src/main/java/org/talend/daikon/avro/visitor/record/VisitableArray.java
@@ -10,11 +10,11 @@
 // 9 rue Pages 92150 Suresnes, France
 //
 // ============================================================================
-package org.talend.daikon.avro.visitor;
+package org.talend.daikon.avro.visitor.record;
 
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
-import org.talend.daikon.avro.path.TraversalPath;
+import org.talend.daikon.avro.visitor.path.TraversalPath;
 
 import java.util.ArrayList;
 import java.util.Collections;
