+++ daikon/src/main/java/org/talend/daikon/avro/visitor/record/VisitableFixed.java
@@ -10,24 +10,22 @@
 // 9 rue Pages 92150 Suresnes, France
 //
 // ============================================================================
-package org.talend.daikon.avro.visitor;
+package org.talend.daikon.avro.visitor.record;
 
-import org.talend.daikon.avro.path.TraversalPath;
-
-import java.nio.ByteBuffer;
+import org.apache.avro.generic.GenericData;
+import org.talend.daikon.avro.visitor.path.TraversalPath;
 
 /**
- * Wrapper implementation for Bytes Avro type
+ * Wrapper implementation for fixed Avro type
  */
-public class VisitableBytes extends AbstractVisitableStructure<ByteBuffer> {
+public class VisitableFixed extends AbstractVisitableStructure<GenericData.Fixed> {
 
-    VisitableBytes(ByteBuffer value, TraversalPath path) {
+    VisitableFixed(GenericData.Fixed value, TraversalPath path) {
         super(value, path);
     }
 
     @Override
     public void accept(RecordVisitor visitor) {
         visitor.visit(this);
-
     }
 }
