+++ daikon/src/main/java/org/talend/daikon/avro/visitor/record/VisitableStructure.java
@@ -10,9 +10,9 @@
 // 9 rue Pages 92150 Suresnes, France
 //
 // ============================================================================
-package org.talend.daikon.avro.visitor;
+package org.talend.daikon.avro.visitor.record;
 
-import org.talend.daikon.avro.path.TraversalPath;
+import org.talend.daikon.avro.visitor.path.TraversalPath;
 
 /**
  * Wrapper for an Avro record field - this is necessary to implement the visitor pattern
