+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/arrayOfArrayOfSimpleType_record.json
