+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/arrayOfArrayOfSimpleType_schema.json
