+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/arrayOfRecords_record.json
