+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/arrayOfSimpleType_schema.json
