+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/mapOfArrays_record.json
