+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/mapOfRecords_record.json
