+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/mapOfSimpleType_record.json
