+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/nestedRecord_record.json
