+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/optionalSimpleTypeNull_record.json
