+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/optionalSimpleTypePresent_record.json
