+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/optionalSimpleType_schema.json
