+++ daikon/src/test/resources/org/talend/daikon/avro/visitor/record/simpleTypes_schema.json
